Tinted statusbar for KitKat
Added tinted statusbar for Android 4.4.x (KitKat).
It's working everywhere except on the settings screen, where I had to
disable it.
As the settings screen doesn't have a layout file I could not set a
white background, meaning it used the WindowBackground which was green
and made the text illegible.